I'd like to enable changes tries in default (staging) chain specification in nearest future so that out next testnet will have them enabled. But beforehand there should be some statistics collected to choose appropriate defaults for `digest_interval` and `digest_levels`. I've tested several configurations already to estimate maximal number of changes that could be covered by top-level digests without significant performance drops. Some rough measurements (that were made without changes in this PR) are at the end of description (fwiw).

What I've found during testing is that there are few easy-to-implement optimizations. This PR cuts block-with-CT (changes trie) import time by ~30% by:
- reducing number of allocations in `build.rs`;
- switching from `HashSet` to `BTreeSet` in `OverlayedValue::extrinsics` because we had to sort this (by converting to `BTreeSet`) anyway;
- instead of computing CT root by calling `trie_root(input_pairs)` and then actually building CT (to store in DB), we're now building CT (the root is computed there anyway).

I have another optimization implemented - in-memory cache that holds data required to build digest tries (instead of reading lower-lever CT from the DB). But: (1) I'm not sure about the case when it actually **significantly** helps - need to confirm that yet (2) this should be a separate PR.

==========================================================================

In the chain where at every block we're changing 1000 different keys:
- importing block without CT takes ~0.01s;
- importing non-digest block with CT takes ~0.018s;
- importing l1-digest block which covers 8 blocks (i.e. there are ~8000 different keys changed within this range) takes ~0.072s;
- importing l2-digest block which covers 64 blocks (i.e. there are ~64000 different keys changed within this range) takes ~0.45s.

In the chain where at every block we're changing 5000 different keys:
- importing block without CT takes ~0.027s;
- importing non-digest block with CT takes ~0.060s;
- importing l1-digest block which covers 8 blocks (i.e. there are ~40000 different keys changed within this range) takes ~0.25s;
- importing l2-digest block which covers 64 blocks (i.e. there are ~320000 different keys changed within this range) takes ~2.106s.