@svyatonik 2019-06-12
Think I found some other things that could be eliminated - inprogress