@svyatonik 2019-06-12
I thought of maybe using the fact that extrinsics are executed in order => maybe replacing `extrinsics: Option<BTreeSet<u32>>` in `OverlayedValue` with `extrinsics: Option<Vec<u32>>` which is naturally ordered, but probably not the best idea, because it is the runtime who updates the `EXTRINSIC_INDEX` => it isn't guaranteed to be in-order.