@svyatonik 2019-06-20
The last update of this PR - I'm adding results for some 'average' case, where every block:
1) updates 130 unique (across all blocks) keys;
2) adds fake updates for 20 unique (across all blocks) keys;
3) updates same 50 keys @ every block;
4) adds fake updates for 10 same keys every block.

With CT enabled, import of block takes ~0.00399s. When digests are created every 512 blocks, it takes ~0.35114s to import that block. When digests are created every 1024 blocks, it takes ~0.69611 to import that block.

As some conclusion, I think that before enabling CT for the testnet:
1) it makes sense to allow change of the CT configuration on active chain;
2) probably later add some auto-maintain mechanism that will limit number of changes that could be covered by digest tries.