@gavofyork 2019-06-21
> limit number of changes that could be covered by digest tries

how would this work?