@svyatonik 2019-06-21
Something like that, but I have no details for this yet: (1) count number of keys that have to be covered by next top-level CT digest (2) when this crosses limit => create digest CT at next block  and emit signal (digest item) that we have created top-level digest CT before schedule (3) reset number of keys to zero and go to (1).

Configuration change is also should be announced to light clients using special digest item.