@svyatonik 2019-06-21
Most probably "number of keys" -> "number of changes tries' input pairs" that should be read to build top level digest block. Because it is impossible to have `len(unique keys from all lower-level digests)` without extra-computations.