@gavofyork 2019-06-25
ok - makes sense. might it be possible to do incremental CT builds to spread the cost over all the blocks rather than have a single block that takes ages?