@svyatonik 2019-06-25
> might it be possible to do incremental CT builds to spread the cost over all the blocks rather than have a single block that takes ages?

I've considered this in the very beginning - i.e. you know that next digest trie will be built at block #1024 => when tries for previous 1023 blocks are being built, you also update trie for block#1024 => when you reach #1024 you actually have an almost-complete digest trie. But then I have thought about forks - i.e. when you have forked at #500 and then fork1 also forks at #700 then you'll have 4 potential future digest tries that have to be updated on every block. It seems too heavy (when thinking about performance) + complex (when thinking about code) to maintain. Dropped idea because of that.